+++ formats/src/main/scala/overflowdb/formats/neo4jcsv/Neo4jCsvExporter.scala
@@ -1,0 +1,12 @@
+package overflowdb.formats.neo4jcsv
+
+import overflowdb.Graph
+import overflowdb.formats.Exporter
+
+object Neo4jCsvExporter extends Exporter {
+
+  override def runExport(graph: Graph, outputRootDirectory: String): Unit = {
+    ???
+  }
+
+}
